Added code to print out the Menu
Doesn't work yet. Committing to test this on my laptop to see if the memory error persists.

+++ Appetiser.cpp
@@ -1,7 +1,7 @@
 #include "Appetiser.h"
 
 // Constructor 
-Appetiser::Appetiser(int icalories, std::string iname, float iprice, bool ishareable, bool itwoForOne) : Item(icalories, iname, iprice)
+Appetiser::Appetiser(int icalories, std::string iname, double iprice, bool ishareable, bool itwoForOne) : Item(icalories, iname, iprice)
 {
 	// First call the parent constructor, then initialise Appetiser class specific parameters
 	shareable = ishareable;
