+++ Appetiser.h
@@ -6,7 +6,7 @@ class Appetiser : public Item
 {
 public:
 	// Constructor
-	Appetiser(int icalories, std::string iname, float iprice, bool ishareable, bool itwoForOne);
+	Appetiser(int icalories, std::string iname, double iprice, bool ishareable, bool itwoForOne);
 	~Appetiser();
 
 	std::string toString();
